fix(ubsan): eliminate three UBSan-strict implicit-conversion violations
pdjson.c: stack_top field was changed to ptrdiff_t in a prior commit but
the sentinel assignments still used (size_t)-1 (UINT64_MAX), which
UBSan flags as an out-of-range implicit conversion to a signed type.
Change both sentinel assignments (init() and json_reset()) to plain -1,
which is well-defined for ptrdiff_t and matches the sentinel comparisons
already present in the file.

vif_avx512.c: two scalar-tail loops in vif_subsample_rd_8_avx512 and
vif_subsample_rd_16_avx512 iterate with unsigned j but compute
int jj = j - fwidth_half without an explicit cast. In C, the signed
operand is promoted to unsigned, causing an implicit conversion from
the resulting unsigned wrap-around value back to int — UBSan-flagged.
Add (int)j casts at both sites to keep the subtraction in signed
arithmetic, matching the pattern already used in the vectorised path
above each scalar tail.

The motion_avx512 scalar-tail overflow (finding #2) was already fixed
in a prior commit with (uint32_t) casts; no change needed there.

+++ core/src/feature/x86/vif_avx512.c
@@ -1420,7 +1420,7 @@ void vif_subsample_rd_8_avx512(VifBuffer buf, unsigned w, unsigned h)
         for (unsigned j = n << 4; j < w; ++j) {
             uint32_t accum_ref = 0;
             uint32_t accum_dis = 0;
-            int jj = j - fwidth_half;
+            int jj = (int)j - fwidth_half;
             int jj_check = jj;
             for (unsigned fj = 0; fj < fwidth; ++fj, jj_check = jj + fj) {
                 const uint16_t fcoeff_scalar = vif_filt_s1[fj];
@@ -1586,7 +1586,7 @@ void vif_subsample_rd_16_avx512(VifBuffer buf, unsigned w, unsigned h, int scale
         for (unsigned j = n << 4; j < w; ++j) {
             uint32_t accum_ref = 0;
             uint32_t accum_dis = 0;
-            int jj = j - fwidth_half;
+            int jj = (int)j - fwidth_half;
             int jj_check = jj;
             for (unsigned fj = 0; fj < fwidth; ++fj, jj_check = jj + fj) {
                 const uint16_t fcoeff = vif_filt[fj];
